chore(router): move SocketPush interface out of models
